Add language code validation to Neo4JDatabaseValidator
Implemented a new method to validate the existence of language codes in the Neo4j database. This method raises an InvalidRequest exception if the provided language code is not found, along with a list of available codes. Updated Neo4JDatabase to utilize this validation when processing requests with a language filter and during expression creation.